backend/requirements.txt
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
python-dotenv==1.0.1
whitenoise==6.7.0
cohere==5.11.2
gunicorn==21.2.0
#pytz
#dj-database-url==2.0.0 